Create Financial_risk.ipynb
In this project, different machine learning models were used to predict the financial risk levels of customers. The best results were achieved with the Random Forest model. This model obtained high accuracy rates by effectively modeling the features of the dataset. In future studies, it may be possible to increase the accuracy of the models by adding more data and features.